// 3205-web/src/shared/api/api.ts
// This file is generated by scripts/gen-api.script.ts. Do not edit.

export const apiUrls = {
  "app": {
    GET_HEALTHCHECK: {
      url: () => `/healthcheck`,
      method: "GET",
      response: "void",
    },
  },
  "job": {
    DELETE_JOB_BY_ID: {
      url: (id: string) => `/api/jobs/${encodeURIComponent(String(id))}`,
      method: "DELETE",
      response: "void",
    },
    GET_API_JOBS: {
      url: () => `/api/jobs`,
      method: "GET",
      response: "GetJobDTO[]",
    },
    GET_JOB_BY_ID: {
      url: (id: string) => `/api/jobs/${encodeURIComponent(String(id))}`,
      method: "GET",
      response: "GetFullJobDTO",
    },
    POST_API_JOBS: {
      url: () => `/api/jobs`,
      method: "POST",
      body: "CreateJobDTO",
      response: "GetJobDTO",
    },
  },
} as const;